requirements.txt
selenium==4.15.0
streamlit==1.28.0
nltk==3.8.1
textblob==0.17.1
pyyaml==6.0.1